refactor(routing): extract router config from App.jsx to dedicated file
