+++ src/App.jsx
@@ -1,213 +1,19 @@
-// App.jsx
-
-import { useState, lazy, Suspense } from "react";
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
+// src/App.jsx
+import { RouterProvider } from "react-router-dom";
 import "./App.css";
 import "./tailwind.css";
-
-// Komponens importok
-import Error from "./pages/Error/Error";
-const Home = lazy(() => import("./pages/Home/Home"));
-const Graph = lazy(() => import("./pages/Graph/Graph"));
-const Test = lazy(() => import("./pages/Test/Test"));
-const Admin = lazy(() => import("./pages/Admin/Admin"));
-const AsteroidePage = lazy(() => import("./pages/Asteroide/Asteroide")); // Győződj meg a névről!
-
-// Layout importok
-const AppLayout = lazy(() => import("./layouts/AppLayout"));
-const LayoutSimple = lazy(() => import("./layouts/LayoutSimple"));
-
-// DataContext és Hook importok
-// A MyDataContext-et most már az AppLayout használja, itt nem feltétlenül kell.
-// import { MyDataContext } from "./context/DataContext";
 import { useSave } from "./hooks/use-saveuser";
-//import useAsteroidData from "./hooks/useAsteroidData";
-import { sentry } from "./locales/nasadummy";
-
-// Fallback komponens
-const LoadingFallback = ({ message = "Loading..." }) => (
-  <div className="w-screen h-screen bg-black text-primary flex items-center justify-center text-xl font-mono">
-    {message}
-  </div>
-);
+import { router } from "./router"; // Importáljuk a kiszervezett routert
 
 function App() {
-  useSave(); // Hook futtatása
+  // A useSave hook itt maradhat, mert ez egy globális,
+  // alkalmazás-szintű mellékhatás.
+  useSave();
 
-  // Dummy adatok
-  //const asteroidData = useAsteroidData();
-  const [sumObject] = useState(sentry);
-  //const defaultVideoSrc = "https://sablonossablon.web.app/video/optimized_earth2.mp4";
-  const defaultVideoSrc = "/muhold_compress.mp4";
-  const router = createBrowserRouter([
-    // --- Minden fő útvonal külön objektum ---
-    {
-      path: "/",
-      element: (
-        <Suspense fallback={<LoadingFallback message="Loading Layout..." />}>
-          <AppLayout
-            showNavbar={false}
-            backgroundConfig={{ type: "video", src: defaultVideoSrc }}
-          />
-        </Suspense>
-      ),
-      errorElement: (
-        <LayoutSimple>
-          <Error />
-        </LayoutSimple>
-      ),
-      children: [
-        // Csak a Home oldal tartozik ide
-        {
-          index: true,
-          element: (
-            <Suspense fallback={<LoadingFallback message="Loading Home..." />}>
-              <Home />
-            </Suspense>
-          ),
-        },
-      ],
-    },
-    {
-      path: "asteroide",
-      element: (
-        <Suspense fallback={<LoadingFallback message="Loading Layout..." />}>
-          {/* Navbar látszik, videó van */}
-          <AppLayout
-            showNavbar={true}
-            backgroundConfig={{
-              type: "image",
-              src: "/images/Asteroid_preview.jpg",
-            }}
-          />
-        </Suspense>
-      ),
-      errorElement: (
-        <LayoutSimple>
-          <Error />
-        </LayoutSimple>
-      ),
-      children: [
-        // Csak az Asteroide oldal tartozik ide
-        {
-          index: true,
-          element: (
-            <Suspense
-              fallback={<LoadingFallback message="Loading Asteroide Page..." />}
-            >
-              <AsteroidePage />
-            </Suspense>
-          ),
-        },
-      ],
-    },
-    {
-      path: "graph",
-      element: (
-        <Suspense fallback={<LoadingFallback message="Loading Layout..." />}>
-          {/* Navbar látszik, videó van */}
-          <AppLayout
-            showNavbar={true}
-            backgroundConfig={{ type: "video", src: defaultVideoSrc }}
-          />
-        </Suspense>
-      ),
-      errorElement: (
-        <LayoutSimple>
-          <Error />
-        </LayoutSimple>
-      ),
-      children: [
-        // Csak a Graph oldal tartozik ide
-        {
-          index: true,
-          element: (
-            <Suspense
-              fallback={<LoadingFallback message="Loading Graph Page..." />}
-            >
-              <Graph sumObject={sumObject} />
-            </Suspense>
-          ),
-        },
-      ],
-    },
-    {
-      path: "test",
-      element: (
-        <Suspense fallback={<LoadingFallback message="Loading Layout..." />}>
-          {/* Navbar látszik, statikus kép van */}
-          <AppLayout
-            showNavbar={true}
-            backgroundConfig={{ type: "image", src: "/images/starfield.png" }}
-          />
-        </Suspense>
-      ),
-      errorElement: (
-        <LayoutSimple>
-          <Error />
-        </LayoutSimple>
-      ),
-      children: [
-        // Csak a Test oldal tartozik ide
-        {
-          index: true,
-          element: (
-            <Suspense
-              fallback={<LoadingFallback message="Loading Test Page..." />}
-            >
-              <Test />
-            </Suspense>
-          ),
-        },
-      ],
-    },
-    {
-      path: "admin",
-      element: (
-        <Suspense fallback={<LoadingFallback message="Loading Layout..." />}>
-          {/* Navbar látszik, nincs extra háttér */}
-          <AppLayout showNavbar={true} backgroundConfig={{ type: "none" }} />
-        </Suspense>
-      ),
-      errorElement: (
-        <LayoutSimple>
-          <Error />
-        </LayoutSimple>
-      ),
-      children: [
-        // Csak az Admin oldal tartozik ide
-        {
-          index: true,
-          element: (
-            <Suspense
-              fallback={<LoadingFallback message="Loading Admin Page..." />}
-            >
-              <Admin />
-            </Suspense>
-          ),
-        },
-      ],
-    },
-    // Catch-all 404 útvonal
-    {
-      path: "*",
-      element: (
-        <Suspense
-          fallback={<LoadingFallback message="Loading Error Page..." />}
-        >
-          <LayoutSimple>
-            <Error />
-          </LayoutSimple>
-        </Suspense>
-      ),
-      // Ennek nincs children-je
-    },
-  ]);
-
-  // Már nincs szükség a külső div-re a témához, mert az AppLayout kezeli
+  // A komponens most már sokkal egyszerűbb.
+  // Csak a RouterProvider-t rendereli.
   return (
     <>
-      {/* A rejtett div a szín kiolvasásához továbbra is kellhet */}
       <div id="theme-color-ref" className="text-primary hidden">
         Sampling div tailwind for react charts for dynamic colors
       </div>
